requirements.txt
torch
nltk
torchtext==0.6.0
pyvi
spacy
PyYAML
dill
pandas
numpy